refactor : 구조 변경 및 정리
