+++ src/main/java/kr/hhplus/ecommerce/application/user/dto/UserCriteria.java
@@ -1,13 +1,15 @@
-//package kr.hhplus.ecommerce.application.user.dto;
-//
-//
-//public record UserCriteria() {
-//
-//    public record Find(
-//            Long id
-//    ){
-//        public static UserCommand.Find toCommand(Long id) {
-//            return new UserCommand.Find(id);
-//        }
-//    }
-//}
+package kr.hhplus.ecommerce.application.user.dto;
+
+
+import kr.hhplus.ecommerce.domain.user.dto.UserCommand;
+
+public record UserCriteria() {
+
+    public record Find(
+            Long id
+    ) {
+        public static UserCommand.Find toCommand(Long id) {
+            return new UserCommand.Find(id);
+        }
+    }
+}
