+++ src/main/java/kr/hhplus/ecommerce/config/jpa/JpaConfig.java
@@ -3,13 +3,11 @@ package kr.hhplus.ecommerce.config.jpa;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.transaction.PlatformTransactionManager;
 
 @Configuration
 @EnableJpaAuditing
-@EnableJpaRepositories
 public class JpaConfig {
 
     @Bean
