+++ src/main/java/kr/hhplus/ecommerce/domain/order/OrderRepository.java
@@ -1,12 +1,10 @@
-package kr.hhplus.ecommerce.domain.order.repository;
+package kr.hhplus.ecommerce.domain.order;
+
 
 import kr.hhplus.ecommerce.domain.order.entity.Order;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-
-@Repository
 public interface OrderRepository {
     Order save(Order order);
