--- src/main/java/kr/hhplus/ecommerce/domain/order/repository/OrderItemRepository.java
@@ -1,16 +1,0 @@
-package kr.hhplus.ecommerce.domain.order.repository;
-
-import kr.hhplus.ecommerce.domain.order.entity.OrderItem;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-
-@Repository
-public interface OrderItemRepository {
-
-    Optional<OrderItem> findByProductOptionId(Long command);
-
-    OrderItem save(OrderItem orderItem);
-
-}
