+++ src/main/java/kr/hhplus/ecommerce/domain/user/UserService.java
@@ -1,9 +1,12 @@
 package kr.hhplus.ecommerce.domain.user;
 
-import kr.hhplus.ecommerce.domain.user.dto.UserInfo;
+import kr.hhplus.ecommerce.config.exception.ErrorCode;
+import kr.hhplus.ecommerce.config.exception.Exception;
+import kr.hhplus.ecommerce.domain.user.dto.UserCommand;
 import kr.hhplus.ecommerce.domain.user.entity.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -11,13 +14,9 @@ public class UserService {
 
     private final UserRepository userRepository;
 
-    public UserInfo.User getUser(Long userId) {
-        User user = userRepository.findById(userId)
-                .orElseThrow(() -> new IllegalArgumentException("사용자가 존재하지 않습니다."));
-
-        return UserInfo.User.builder()
-                .userId(user.getId())
-                .username(user.getName())
-                .build();
+    @Transactional(readOnly = true)
+    public User findByUserId(UserCommand.Find command) {
+        return userRepository.findById(command.id())
+                .orElseThrow(() -> new Exception(ErrorCode.NOT_FOUND));
     }
 }
