+++ src/main/java/kr/hhplus/ecommerce/domain/user/dto/UserCommand.java
@@ -1,0 +1,12 @@
+package kr.hhplus.ecommerce.domain.user.dto;
+
+public record UserCommand() {
+
+    public record Find(
+            Long id
+    ) {
+        public static Find from(Long id) {
+            return new Find(id);
+        }
+    }
+}
