+++ src/test/java/kr/hhplus/ecommerce/interfaces/ControllerCommonTest.java
@@ -1,20 +1,14 @@
 package kr.hhplus.ecommerce.interfaces;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import kr.hhplus.ecommerce.interfaces.order.OrderController;
+import kr.hhplus.ecommerce.application.point.PointFacade;
 import kr.hhplus.ecommerce.interfaces.point.PointController;
-import kr.hhplus.ecommerce.interfaces.presentation.controller.*;
-import kr.hhplus.ecommerce.interfaces.user.CouponController;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.web.servlet.MockMvc;
 
 @WebMvcTest(controllers = {
-    PointController.class,
-        CouponController.class,
-        OrderController.class,
-        StatController.class
-
+        PointController.class
 })
 public abstract class ControllerCommonTest {
 
@@ -23,4 +17,19 @@ public abstract class ControllerCommonTest {
 
     @Autowired
     protected ObjectMapper objectMapper;
+
+    @Autowired
+    protected PointFacade pointFacade;
+
+//    @MockitoBean
+//    protected ProductService productService;
+
+//    @MockitoBean
+//    protected BalanceFacade balanceFacade;
+//
+//    @MockitoBean
+//    protected OrderFacade orderFacade;
+//
+//    @MockitoBean
+//    protected UserCouponService userCouponService;
 }
